move bySeverity to utils
